--- src/abigen/generator.js
@@ -1,2 +1,0 @@
-// Converted to CJS file `generator.cjs`; this file intentionally left as a thin wrapper for ESM builds if needed.
-module.exports = require("./generator.cjs");
