--- src/abigen/wasm/wasm_exec_node.js
@@ -1,2 +1,0 @@
-// ESM wrapper that imports the CJS initializer for Node environment
-import "./wasm_exec_node.cjs";
