+++ test/fixture-projects/hardhat-project-defined-config/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "hardhat-project-defined-config",
+  "version": "0.1.0",
+  "license": "MIT",
+  "author": "Distributed Lab",
+  "readme": "README.md",
+  "type": "module"
+}
